# radical_backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BOOK APP API",
        "description": "Book List Api",
        "version": "1.0"
    },
    "paths": {
        "/api/book": {
            "get": {
                "tags": [
                    "Book"
                ],
                "summary": "Get List of books",
                "operationId": "0e75dde625f5fdef923b42b410113da4",
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "author": {
                                                        "type": "string",
                                                        "example": "example name"
                                                    },
                                                    "price": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": "example price"
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "example title"
                                                    },
                                                    "rate": {
                                                        "type": "number",
                                                        "example": "example rate"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "example description"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2021-12-11T09:25:53.000000Z"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2021-12-11T09:25:53.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Book"
                ],
                "summary": "Create book",
                "operationId": "d186ef540836dc71b8d1342642a1cce5",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "author": {
                                                "type": "string"
                                            },
                                            "price": {
                                                "type": "number",
                                                "format": "float"
                                            },
                                            "title": {
                                                "type": "string"
                                            },
                                            "rate": {
                                                "type": "number"
                                            },
                                            "description": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "author": "example name",
                                    "price": "example price",
                                    "title": "example title",
                                    "rate": "example rate",
                                    "description": "example description"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "author": {
                                            "type": "string",
                                            "example": "name"
                                        },
                                        "price": {
                                            "type": "number",
                                            "format": "float",
                                            "example": "price"
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "title"
                                        },
                                        "rate": {
                                            "type": "number",
                                            "example": "rate"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "description"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "fail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/book/{id}": {
            "get": {
                "tags": [
                    "Book"
                ],
                "summary": "Get a book",
                "operationId": "c2d58ab5f9eb729975ff3468e0e327ac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "author": {
                                            "type": "string",
                                            "example": "name"
                                        },
                                        "price": {
                                            "type": "number",
                                            "format": "float",
                                            "example": "price"
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "title"
                                        },
                                        "rate": {
                                            "type": "number",
                                            "example": "rate"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "description"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Book"
                ],
                "summary": "Update book",
                "operationId": "a0894639dd8663ef8249b56ed6fe561a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "author": {
                                                "type": "string"
                                            },
                                            "price": {
                                                "type": "number",
                                                "format": "float"
                                            },
                                            "title": {
                                                "type": "string"
                                            },
                                            "rate": {
                                                "type": "number"
                                            },
                                            "description": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "author": "example name",
                                    "price": "example price",
                                    "title": "example title",
                                    "rate": "example rate",
                                    "description": "exampledescription"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "numberr",
                                            "example": 1
                                        },
                                        "author": {
                                            "type": "string",
                                            "example": "name"
                                        },
                                        "price": {
                                            "type": "number",
                                            "format": "float",
                                            "example": "price"
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "title"
                                        },
                                        "rate": {
                                            "type": "number",
                                            "example": "rate"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "description"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Book"
                ],
                "summary": "Delete book",
                "operationId": "e1d27bcf7c73344045c09c19499d956d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "THE BOOK IS DELETED"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Book",
            "description": "Book"
        }
    ]
}